Feedback from Fabien, fix legacy error messages
