+++ spec/client/config_spec.rb
@@ -33,26 +33,26 @@ RSpec.describe Client::Config do # rubocop:disable Metrics/BlockLength
         expect do
           config.validate!
         end.to raise_error(NameError,
-                           "Error: uninitialized constant APP_ID in .env, Error: uninitialized constant PRIVATE_KEY in"\
-                           " .env, Error: uninitialized constant ROOT_PATH_TO_SAVE in .env")
+                           "Error: uninitialized configuration app_id, Error: uninitialized configuration private_key,"\
+                           " Error: uninitialized configuration root_path_to_save")
       end
 
       it "raises a NameError when app_id is missing" do
         config.private_key = "test_private_key"
         config.root_path_to_save = "/path/to/save"
-        expect { config.validate! }.to raise_error(NameError, "Error: uninitialized constant APP_ID in .env")
+        expect { config.validate! }.to raise_error(NameError, "Error: uninitialized configuration app_id")
       end
 
       it "raises a NameError when private_key is missing" do
         config.app_id = "test_app_id"
         config.root_path_to_save = "/path/to/save"
-        expect { config.validate! }.to raise_error(NameError, "Error: uninitialized constant PRIVATE_KEY in .env")
+        expect { config.validate! }.to raise_error(NameError, "Error: uninitialized configuration private_key")
       end
 
       it "raises a NameError when root_path_to_save is missing" do
         config.app_id = "test_app_id"
         config.private_key = "test_private_key"
-        expect { config.validate! }.to raise_error(NameError, "Error: uninitialized constant ROOT_PATH_TO_SAVE in .env")
+        expect { config.validate! }.to raise_error(NameError, "Error: uninitialized configuration root_path_to_save")
       end
     end
   end
